fix and adapt code to arduino ide
