+++ Embebido/src/main/setup_utils.h
@@ -12,8 +12,8 @@
 #define TIMEOUT_GET_TIME 5
 #define DEBOUNCE_SECONDS 200
 
-const char *ssid = "Wokwi-GUEST";
-const char *password = "";
+const char *ssid = "CLAROVLG";
+const char *password = "F6BCA8C4";
 
 const char *ntpServer = "pool.ntp.org";
 const long gmtOffset_sec = GMT_DIFFERENCE * HOUR_TO_SECONDS;
@@ -117,8 +117,8 @@ void detectButtonPress()
 
 void queueSetup()
 {
- timeEventsQueue = xQueueCreate(MAX_EVENTS_QUEUE, sizeof(events));
- buttonEventsQueue = xQueueCreate(MAX_EVENTS_QUEUE, sizeof(short));
+ timeEventsQueue = xQueueCreate(MAX_EVENTS, sizeof(events));
+ buttonEventsQueue = xQueueCreate(MAX_EVENTS, sizeof(short));
 }
 
 void semaphoreSetup()
