+++ Embebido/src/main/state_machine.h
