+++ Embebido/src/main/state_machine_actions.h
